+++ configs/vite.config.base.ts
