+++ configs/vitest.config.base.browser.ts
