+++ packages/beacon-node/vitest.spec.config.ts
@@ -1,6 +1,6 @@
 import {defineConfig, mergeConfig} from "vitest/config";
 import {buildTargetPlugin} from "../../scripts/vite/plugins/buildTargetPlugin";
-import vitestConfig from "../../configs/vitest.base.spec.config";
+import vitestConfig from "../../configs/vitest.config.base.spec";
 
 export default mergeConfig(
   vitestConfig,
