+++ packages/params/vitest.browser.config.ts
@@ -1,5 +1,5 @@
 import {defineConfig, mergeConfig} from "vitest/config";
-import vitestConfig from "../../configs/vitest.base.browser.config";
+import vitestConfig from "../../configs/vitest.config.base.browser";
 
 export default mergeConfig(
   vitestConfig,
