+++ packages/spec-test-util/vitest.e2e.config.ts
@@ -1,5 +1,5 @@
 import {defineConfig, mergeConfig} from "vitest/config";
-import vitestConfig from "../../configs/vitest.base.e2e.config";
+import vitestConfig from "../../configs/vitest.config.base.e2e";
 
 export default mergeConfig(
   vitestConfig,
