+++ packages/utils/vitest.config.ts
@@ -1,5 +1,5 @@
 import {defineConfig, mergeConfig} from "vitest/config";
-import vitestConfig from "../../configs/vitest.base.unit.config";
+import vitestConfig from "../../configs/vitest.config.base.unit";
 
 export default mergeConfig(
   vitestConfig,
